Updated resource claims
When creating a resource claim, it will now try to fill in the blanks to ensure a consistent structure in the crowdsourcing task structure. If it's a canvas without a manifest or collection it will first look up all of the manifests that the canvas is in (in the same project) and if there's only one, it will use that. The same process for collections. This ensures that 99% of the time there is only one set of crowdsourcing tasks created in the UI.